avance de middleware

+++ app/Http/Controllers/LogisticController.php
@@ -8,11 +8,17 @@ use Illuminate\Support\Facades\Auth;
 
 class LogisticController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth');
+    }
     /**
      * Display a listing of the resource.
      */
     public function index()
     {
+        $response = Controller::authRoles(2);
+        if ($response) {return $response;}
         $logistics = Logistic::all();
         $user = Auth::user();
         return view('logistics.index', compact('logistics', 'user'));
@@ -32,12 +38,14 @@ class LogisticController extends Controller
      */
     public function store(Request $request)
     {
+        $user = Auth::user();
         $logistic = new Logistic();
-        $logistic->name = $request->name;
+        $logistic->celular = $request->celular;
         $logistic->description = $request->description;
+        $logistic->id_users = $user->id;
         $logistic->save();
 
-        return redirect(route('logistics.index'));
+        return redirect(route('home'));
     }
 
     /**
@@ -45,6 +53,8 @@ class LogisticController extends Controller
      */
     public function show($id)
     {
+        $response = Controller::authRoles(2);
+        if ($response) {return $response;}
         $logistic = Logistic::find($id);
         $user = Auth::user();
         return view('logistics.show', compact('logistic', 'user'));
@@ -55,6 +65,8 @@ class LogisticController extends Controller
      */
     public function edit(string $id)
     {
+        $response = Controller::authRoles(2);
+        if ($response) {return $response;}
         $logistic = Logistic::find($id);
         $user = Auth::user();
         return view('logistics.edit', compact('logistic', 'user'));
@@ -65,6 +77,8 @@ class LogisticController extends Controller
      */
     public function update(Request $request, string $id)
     {
+        $response = Controller::authRoles(2);
+        if ($response) {return $response;}
         $logistic = Logistic::find($id);
         $logistic->name = $request->name;
         $logistic->description = $request->description;
@@ -78,9 +92,10 @@ class LogisticController extends Controller
      */
     public function destroy(string $id)
     {
+        $response = Controller::authRoles(2);
+        if ($response) {return $response;}
         $logistic = Logistic::find($id);
         $logistic->delete();
-
         return redirect(route('logistics.index'));
     }
 }
